Fix return/continue bug, updated readme re model.json, update isObject logic in merge function
