+++ geoSegment/utils/readJsonNullIfNotExists.js
@@ -1,0 +1,17 @@
+/**
+ * @param {string} filePath
+ */
+async function readJsonNullIfNotExists(filePath) {
+  try {
+    return await fs.readJson(filePath);
+  } catch (error) {
+    if (error && error.code === 'ENOENT') {
+      return null;
+    }
+    throw error;
+  }
+}
+
+module.exports = {
+  readJsonNullIfNotExists,
+};
