+++ package.json
@@ -33,6 +33,7 @@
     "fs-extra": "^9.0.1",
     "geolib": "^3.3.1",
     "joi": "^17.2.1",
+    "lodash.isobject": "^3.0.2",
     "moment-timezone": "^0.5.31"
   },
   "engines": {
